Add new symbolic locations
LOCATION_STZONE (the Spell & Trap zone, so excluding mr3 pzones and field zone), LOCATION_MMZONE and LOCATION_EMZONE. Usable in GetMatchingCard, GetFieldGroup, etc

+++ card.h
@@ -59,7 +59,7 @@ struct card_state {
 	card* reason_card{};
 	uint8_t reason_player{};
 	effect* reason_effect{};
-	bool is_location(int32_t loc) const;
+	bool is_location(uint16_t loc) const;
 	void set0xff();
 };
